+++ src/js/features/progress.js
@@ -4,7 +4,7 @@ import document from 'global/document';
 import {config} from '../player';
 import MediaElementPlayer from '../player';
 import i18n from '../core/i18n';
-import {IS_FIREFOX, IS_IOS, IS_ANDROID} from '../utils/constants';
+import {IS_FIREFOX, IS_IOS, IS_ANDROID, SUPPORT_PASSIVE_EVENT} from '../utils/constants';
 import {secondsToTimeCode} from '../utils/time';
 import {offset, addClass, removeClass, hasClass} from '../utils/dom';
 
@@ -447,7 +447,7 @@ Object.assign(MediaElementPlayer.prototype, {
 						});
 					}
 				}
-			});
+			}, (SUPPORT_PASSIVE_EVENT && events[i] === 'touchstart') ? { passive: true } : false);
 		}
 		t.slider.addEventListener('mouseenter', (e) => {
 			if (e.target === t.slider && t.getDuration() !== Infinity) {